# Tweak the Batch.*Deferred implementation for performance