Rather than returning a `DeferredBatchOp`, the `Batch.*Deferred` methods
now return a `*DeferredBatchOp`. This replaces copying the
`DeferredBatchOp` struct (64 bytes) with a pointer (8 bytes). This sped
up the benchmarks sufficiently that we can use the `*Deferred` methods
to implement the non-deferred methods.

Add a `Batch.count` field, rather than incrementing the count directly
in the batch representation. This is both a minor performance
win (`Batch.increment` showed up in the `BatchSet` benchmark), as well
as a convenience win `Batch.Reset` no longer mutates the value returned
by `Batch.Repr()`. There were a handful of places in the code which were
previously accessing or setting `Batch.storage.data` directly which were
changed to use `{Repr,SetRepr}()`.

````
name                       old time/op  new time/op  delta
BatchSet-8                 29.4ns ± 6%  27.9ns ± 4%   -4.97%  (p=0.000 n=10+10)
IndexedBatchSet-8           225ns ± 7%   231ns ±11%     ~     (p=0.324 n=9+10)
BatchSetDeferred-8         38.1ns ± 8%  26.9ns ± 2%  -29.43%  (p=0.000 n=10+9)
IndexedBatchSetDeferred-8   230ns ± 9%   229ns ± 5%     ~     (p=0.958 n=10+10)
```